[Refactor] Divide actions into multiple files
