+++ src/redux/actions/categoryActions.js
@@ -1,0 +1,26 @@
+export function setCategories(categories) {
+  return {
+    type: 'setCategories',
+    payload: {
+      categories,
+    },
+  };
+}
+
+export function updateCategoriesLoading(loading) {
+  return {
+    type: 'updateCategoriesLoading',
+    payload: {
+      loading,
+    },
+  };
+}
+
+export function selectCategory(categoryId) {
+  return {
+    type: 'selectCategory',
+    payload: {
+      categoryId,
+    },
+  };
+}
