+++ src/redux/actions/index.js
@@ -1,0 +1,4 @@
+export * from './regionActions';
+export * from './categoryActions';
+export * from './restaurantActions';
+export * from './asyncActions';
