+++ src/redux/actions/restaurantActions.js
@@ -1,0 +1,13 @@
+export function updateRestaurantsLoading(loading) {
+  return {
+    type: 'updateRestaurantsLoading',
+    payload: {
+      loading,
+    },
+  };
+}
+
+// TODO: delete this
+export function xxo() {
+
+}
