+++ src/redux/store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 
 import thunk from 'redux-thunk';
 
-import reducer from './reducers/index';
+import reducer from './reducers';
 
 const store = createStore(reducer, applyMiddleware(thunk));
